fix: avoid i32 overflow in modular XYB -> f32 conversion (jxl-rs #833)
ConvertModularXYBToF32Stage computed (b + y) in i32 before converting;
with attacker-controlled samples the sum overflows. Found by our own
ClusterFuzzLite batch run 32539550999 (2026-08-22, decode_with_limits,
'attempt to add with overflow' at convert.rs:114); upstream fixed the
same line in #833. Convert each operand to f32 first.

Regression seeds added to fuzz/regression/ (replayed by
tests/fuzz_regression.rs; the overflow is a panic under debug/fuzz
builds and a silent wrap in release):
- modular-xyb-convert-add-overflow-833: the 188-byte cflite crash input
- jxlrs-issue82{3a,3b,4,5}-poc: the upstream July fuzzer cases
  (jxl-rs #823/#824/#825). They came from a transform-engine rewrite the
  fork does not have and are rejected cleanly today; kept as guards.

+++ zenjxl-decoder/src/render/stages/convert.rs
@@ -111,7 +111,9 @@ impl RenderPipelineInOutStage for ConvertModularXYBToF32Stage {
         for i in 0..xsize {
             output_x[0][i] = input_x[0][i] as f32 * scale_x;
             output_y[0][i] = input_y[0][i] as f32 * scale_y;
-            output_b[0][i] = (input_b[0][i] + input_y[0][i]) as f32 * scale_b;
+            // Add in f32: the two i32 samples are attacker-controlled and their
+            // sum can exceed i32::MAX (found by ClusterFuzzLite; jxl-rs #833).
+            output_b[0][i] = (input_b[0][i] as f32 + input_y[0][i] as f32) * scale_b;
         }
     }
 }
